refactor mps tests
