+++ docs/source/mps/measurements.rst
@@ -15,7 +15,7 @@ by contracting a network formed by MPS and conjugate of another or the same MPS.
 While the above allows calculating the norm of MPS or MPO, a dedicated method extracts the norm through canonization.
 This can be more precise for norms close to zero.
 
-.. autofunction:: yastn.tn.mps.MpsMpo.norm
+.. automethod:: yastn.tn.mps.MpsMpo.norm
 
 :ref:`Measurement<theory/mps/basics:Measurements>` of MPO's expectation value
 is calculated by contracting a network formed by MPS, MPO, and a conjugate
@@ -34,7 +34,6 @@ One- and two-point expectation values
 -------------------------------------
 
 .. autofunction:: yastn.tn.mps.measure_1site
-
 .. autofunction:: yastn.tn.mps.measure_2site
 
 
@@ -46,6 +45,5 @@ each of the bonds. This amounts to SVD decomposition with respect to a bond,
 where all sites to the left are in :ref:`left-canonical form<theory/mps/basics:canonical form>` and all sites to the right
 are in :ref:`right-canonical form<theory/mps/basics:canonical form>`.
 
-.. autofunction:: yastn.tn.mps.MpsMpo.get_Schmidt_values
-
-.. autofunction:: yastn.tn.mps.MpsMpo.get_entropy
+.. automethod:: yastn.tn.mps.MpsMpo.get_Schmidt_values
+.. automethod:: yastn.tn.mps.MpsMpo.get_entropy
